renamed main page to index.php, moved all functions out of index into functions file and moved that to src directory. Functions files required into index.php. Type hints inclued, some repetitions removed.≈ removed
